requirements.txt
pandas
openpyxl
supabase
python-dotenv